+++ univeq/jvm/src/main/scala/japgolly/univeq/PlatformUnivEq.scala
@@ -1,0 +1,5 @@
+package japgolly.univeq
+
+trait PlatformUnivEq {
+
+}